backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
httpx>=0.28.1
scikit-learn==1.4.2
pandas==2.2.1
numpy==1.26.4
joblib==1.4.0
aiosqlite==0.20.0
openai>=1.52.0
python-jose[cryptography]
passlib[bcrypt]
google-generativeai>=0.8.0
openai-whisper>=20231117
python-multipart>=0.0.9
